LUCENE-4514: Make Scorer.freq() well defined: number of matches in doc
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/branch_4x@1403789 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/core/src/java/org/apache/lucene/search/BooleanScorer.java
@@ -127,7 +127,7 @@ final class BooleanScorer extends Scorer {
     public int docID() { return doc; }
 
     @Override
-    public float freq() { return freq; }
+    public int freq() { return freq; }
 
     @Override
     public int nextDoc() { return NO_MORE_DOCS; }
@@ -322,7 +322,7 @@ final class BooleanScorer extends Scorer {
   }
 
   @Override
-  public float freq() throws IOException {
+  public int freq() throws IOException {
     throw new UnsupportedOperationException();
   }
